u: `SimulationSettings`를 store 도메인으로 이동
