+++ app/store/status.ts
@@ -1,9 +1,3 @@
-import type {
-  RunningStatus,
-  SimulationDifficulty,
-  SimulationSettings,
-} from "~/engine/settings";
-
 interface CoreStatus {
   status: "busy" | "idle";
 }
@@ -29,6 +23,36 @@ export interface TemporalStatus {
   setSuccessRequest: (successRequest: number) => void;
   setNodes: (nodes: NodeMetrics[]) => void;
   setLinks: (links: LinkMetrics[]) => void;
+}
+
+export const SIMULATION_DIFFICULTY = {
+  EASY: "easy",
+  NORMAL: "normal",
+  HARD: "hard",
+};
+export const RUNNING_STATUS = {
+  RUNNING: "running",
+  PAUSED: "paused",
+  STOPPED: "stopped",
+};
+
+export type SimulationDifficulty =
+  (typeof SIMULATION_DIFFICULTY)[keyof typeof SIMULATION_DIFFICULTY];
+export type RunningStatus =
+  (typeof RUNNING_STATUS)[keyof typeof RUNNING_STATUS];
+
+export interface NodeSpec {
+  id: string;
+  coreCount: number;
+}
+
+export interface SimulationSettings {
+  runningStatus: RunningStatus;
+  simulationScale: number;
+  difficulty: SimulationDifficulty;
+  totalRequest: number;
+  timeLimit: number;
+  nodes: NodeSpec[];
 }
 
 export interface SimulationConfig extends SimulationSettings {
